requirements.txt
apache-airflow==2.7.1
python-dotenv==1.0.0
pendulum==2.1.2
sqlalchemy<2.0.0
flask-session==0.5.0
connexion[flask]==2.14.2